requirements.txt
torch==1.4.0
torchvision==0.5.0
lmdb
tqdm
easydict
cupy-cuda100==7.8.0
opencv-python
imageio
natsort
cython
pyyaml
matplotlib
munkres
tensorboardx
terminaltables
timm
visdom
imutils
numpy
scipy
anycurve
yacs
moviepy
h5py
PyTurboJPEG